+++ CHANGELOG.md
@@ -2,13 +2,13 @@
 
 ## Unreleased changes
 
-- `Moar.Duration.to_short_string/1` converts duration to short strings like `"3s"`.
 - Add `Moar.Duration.humanize/1` which can turn `{120, :second}` into `{2, :minute}`,
   and `Moar.Duration.shift/1` which is like `Moar.Duration.convert/1` but returns a duration tuple.
 - Add `Moar.Duration.ago/1` which returns the duration between a given datetime and now.
 - Add `Moar.Duration.between/2` which returns the duration between two datetimes.
 - Add `Moar.Duration.approx/1` which shifts a duration to a simpler approximate value.
 - `Moar.Duration` supports more time units: `:approx_month` (30 days), and `approx_year` (12 approx_months).
+- Add `Moar.Duration.format/2` which formats a duration in either long (`"3 seconds"`) or short (`"3s"`) format.
 
 ## 1.11.0